Move template to components
