+++ src/pages/landing.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Template from "./template";
+import Template from "../components/Template/template";
 
 const Landing = () => <Template />;
